requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
numpy==2.1.1
pydantic==2.8.2
joblib==1.4.2
psycopg2-binary==2.9.9
